refactor(exec): move MoeFfnContext to exec/moe_ffn_context.h
